+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Add a dedicated mode for editing Joker code. (`clojure-ts-joker-mode`).
 - [#113](https://github.com/clojure-emacs/clojure-ts-mode/pull/113): Fix non-working refactoring commands for Emacs-30.
+- [#114](https://github.com/clojure-emacs/clojure-ts-mode/pull/114): Extend built-in completion to complete keywords and local bindings in
+  `for` and `doseq` forms.
 
 ## 0.5.1 (2025-06-17)